@@ -1,6 +1,6 @@
 package com.systemcontrol.corpsele.systemcontrol;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.media.*;
 import android.content.*;
